Move PluginItem code to a separate Item subfolder (#28)
PluginItem is one of several Item types that can go into the Grid. So make room for other types in the folder structure
