+++ src/Item/PluginItem/Interpretations/Interpretations.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import Interpretation from './Interpretation';
 import InputField from './InputField';
 import { tGetInterpretations, tPostInterpretation } from './actions';
-import * as fromReducers from '../../reducers';
+import * as fromReducers from '../../../reducers';
 import { colors } from '../colors';
 
 const style = {
